requirements.txt
torch
torchvision
torchaudio
tensorboardX
tqdm
scikit-learn
PyYAML